+++ Include/VaLib/Utils/Iter.hpp
@@ -1,0 +1,13 @@
+// VaLib - Vast Library
+// Licensed under GNU GPL v3 License. See LICENSE file.
+// (C) 2025 VaLibTeam
+#pragma once
+
+template <typename T>
+class ReverseIter {
+  protected:
+    T iter;
+
+  public:
+    ReverseIter
+};