+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - Refactored build.sh and added new functions like building archives and improved flags support.
 - Added CHANGELOG.md, and scripts/new.py.
+- Fix '--build-all' in build.sh and --default, --default-devel flags.
 ### Changed
 - Updated Doxyfile, for displaying README.md as main page.
 ### Fixed